Make the proper data types in CacheManager

+++ src/components/policy/include/policy/cache_manager.h
@@ -98,7 +98,7 @@ class CacheManager : public CacheManagerInterface {
    * @param current value in days after epoch
    * @return value in days
    */
-  virtual int DaysBeforeExchange(int current);
+  virtual int DaysBeforeExchange(uint16_t current);
 
   /**
    * @brief Increment number of ignition cycles since last exchange by 1
